Move unsharded-LoRA shape check into get_peft_model_state_dict
Address review on #3251:
- relocate the lora_A/lora_B shape validation from PeftModel.save_pretrained
  into get_peft_model_state_dict, next to the existing method-level handling
- only run it for internally-inferred state dicts (skip when an explicit
  state_dict is passed, since its provenance is unknown)
- collapse tests to test_save_pretrained_raises_for_unsharded_state_dict and
  test_save_pretrained_succeeds_for_normal_lora in test_low_level_api.py,
  driving the broken case by overwriting lora_A with a flat tensor
- rename class to TestStateDictFsdpShapeValidation; reword the mitigation note
