+++ tests/test_initialization.py
@@ -3133,91 +3133,6 @@ class TestNoInfiniteRecursionDeepspeed:
             cls.__init__ = original_init
 
 
-class TestSaveValidatesLoraShapes:
-    # Regression for the silent-corruption failure mode described in
-    # vllm-project/vllm#28640 and huggingface/transformers#45313: when PEFT runs save_pretrained() under
-    # DeepSpeed ZeRO-3 / FSDP without first gathering shards, lora_A / lora_B end up 1-D or shape (0,) on disk.
-    # The artifact looks valid but breaks downstream loaders such as vLLM hot-swap with a confusing IndexError.
-    # _validate_lora_adapter_state_dict raises at write time with an actionable hint instead.
-
-    def test_validator_accepts_well_formed_state_dict(self):
-        from peft.utils.save_and_load import _validate_lora_adapter_state_dict
-
-        good = {
-            "base_model.model.q_proj.lora_A.weight": torch.zeros(8, 64),
-            "base_model.model.q_proj.lora_B.weight": torch.zeros(64, 8),
-            # DoRA's magnitude vector is legitimately 1-D and should NOT be flagged.
-            "base_model.model.q_proj.lora_magnitude_vector": torch.zeros(64),
-            # Non-LoRA keys are ignored.
-            "base_model.model.q_proj.bias": torch.zeros(64),
-        }
-        # Must not raise.
-        _validate_lora_adapter_state_dict(good, adapter_name="default")
-
-    @pytest.mark.parametrize(
-        "bad_shape",
-        [
-            (0,),  # safetensors-emitted empty tensor (Z3 rank not owning the slice)
-            (8,),  # 1-D rank-local shard
-        ],
-    )
-    def test_validator_rejects_unsharded_lora_a(self, bad_shape):
-        from peft.utils.save_and_load import _validate_lora_adapter_state_dict
-
-        bad = {
-            "base_model.model.q_proj.lora_A.weight": torch.zeros(*bad_shape),
-            "base_model.model.q_proj.lora_B.weight": torch.zeros(64, 8),
-        }
-        with pytest.raises(ValueError, match=r"DeepSpeed ZeRO-3 / FSDP shards"):
-            _validate_lora_adapter_state_dict(bad, adapter_name="default")
-
-    def test_validator_rejects_zero_sized_lora_b(self):
-        from peft.utils.save_and_load import _validate_lora_adapter_state_dict
-
-        bad = {
-            "base_model.model.q_proj.lora_A.weight": torch.zeros(8, 64),
-            "base_model.model.q_proj.lora_B.weight": torch.zeros(0, 8),
-        }
-        with pytest.raises(ValueError, match=r"lora_B"):
-            _validate_lora_adapter_state_dict(bad, adapter_name="default")
-
-    def test_validator_error_includes_adapter_name(self):
-        from peft.utils.save_and_load import _validate_lora_adapter_state_dict
-
-        bad = {"base_model.model.q_proj.lora_A.weight": torch.zeros(8)}
-        with pytest.raises(ValueError, match=r"'my-adapter'"):
-            _validate_lora_adapter_state_dict(bad, adapter_name="my-adapter")
-
-    def test_save_pretrained_raises_for_unsharded_state_dict(self, tmp_path):
-        # End-to-end: feeding a deliberately-broken state_dict into save_pretrained's kwargs path triggers the
-        # validator at the same call site real Z3 / FSDP saves would.
-        class _Tiny(nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.linear = nn.Linear(8, 8)
-
-        peft_model = get_peft_model(_Tiny(), LoraConfig(target_modules=["linear"], r=4))
-
-        # Include both legitimate (kept) and broken (lora_A 1-D) keys; the broken one should trip the validator.
-        broken_state_dict = {
-            "base_model.model.linear.lora_A.default.weight": torch.zeros(0),
-            "base_model.model.linear.lora_B.default.weight": torch.zeros(8, 4),
-        }
-        with pytest.raises(ValueError, match=r"DeepSpeed ZeRO-3 / FSDP shards"):
-            peft_model.save_pretrained(tmp_path / "broken-adapter", state_dict=broken_state_dict)
-
-    def test_save_pretrained_succeeds_for_normal_lora(self, tmp_path):
-        # Sanity: the validator does NOT regress the happy path.
-        class _Tiny(nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.linear = nn.Linear(8, 8)
-
-        peft_model = get_peft_model(_Tiny(), LoraConfig(target_modules=["linear"], r=4))
-        peft_model.save_pretrained(tmp_path / "ok-adapter")
-        assert (tmp_path / "ok-adapter" / "adapter_config.json").exists()
-
-
 class TestLoadAdapterOfflineMode:
     base_model = "peft-internal-testing/tiny-random-OPTForCausalLM"
     peft_model_id = "peft-internal-testing/tiny-OPTForCausalLM-lora"
